refactor: avoid repeating the return type from the declaration

+++ src/frontend/token.hpp
@@ -37,7 +37,7 @@ struct Token {
     int line;
     int column;
     
-    Token(TokenType t, const std::string& lex, int ln, int col)
+    Token(TokenType t, const std::string& lex, size_t ln, size_t col)
         : type(t), lexeme(lex), line(ln), column(col) {}
 };
 }
